+++ pom.xml
@@ -189,5 +189,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
